+++ custom_components/vivertine/const.py
@@ -64,6 +64,7 @@ DATA_TODAYS_CLASSES = "todays_classes"
 DATA_NEXT_CLASS = "next_class"
 DATA_NEXT_FAVORITE_CLASS = "next_favorite_class"
 DATA_NEXT_FAVORITE_INSTRUCTOR_CLASS = "next_favorite_instructor_class"
+DATA_RECOMMENDED_CLASS = "recommended_class"
 DATA_WEEKLY_VISITS = "weekly_visits"
 DATA_MONTHLY_VISITS = "monthly_visits"
 
@@ -86,6 +87,7 @@ SENSOR_TOTAL_VISITS = "total_visits"
 SENSOR_ACTIVE_BOOKINGS = "active_bookings"
 SENSOR_NEXT_FAVORITE_CLASS = "next_favorite_class"
 SENSOR_NEXT_FAVORITE_INSTRUCTOR_CLASS = "next_favorite_instructor_class"
+SENSOR_RECOMMENDED_CLASS = "recommended_class"
 
 SENSOR_TYPES = {
     SENSOR_MEMBERSHIP_STATUS: {
@@ -157,6 +159,12 @@ SENSOR_TYPES = {
     SENSOR_NEXT_FAVORITE_INSTRUCTOR_CLASS: {
         "name": "Next Favorite Instructor Class",
         "icon": "mdi:account-heart",
+        "unit": None,
+        "device_class": None,
+    },
+    SENSOR_RECOMMENDED_CLASS: {
+        "name": "Recommended Class",
+        "icon": "mdi:star",
         "unit": None,
         "device_class": None,
     },